Update AList.h
Homework assignmern A1: For the assignment, our implementation will be a C++ class called AList.  We will think of the list of items in an AList instance as having left and right ends, so the basic operations will be addLeft, addRight, deleteLeft and deleteRight.   So, for example, if L is an instance of the AList class, and the current list contents of L are <1,2,3>, then after executing the instruction L.addLeft(4), the list contents will be <4,1,2,3>.  If we then execute L.deleteRight(), the list contents will be <4,1,2>.